- refactor: AdminRunner 분리하여 순환 참조 문제 방지
